backend/requirements.txt
Flask
mysql-connector-python==8.1.0
flask_cors
ddtrace